requirements.txt
# Production dependencies for Jira integration
jira>=3.5.0
python-dotenv>=1.0.0
requests>=2.31.0
